Cleaned up TeamsAgentExtension
